Move CapabilityHelper to new helpers module
